embedding/requirements.txt
pymongo
python-dotenv
openai
sentence-transformers
pymongoarrow
tiktoken

